chore: improve handling of filtering on expenses insights
