+++ src/components/transactions/transations.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react"
+import { useEffect, useState, useTransition as useTransitionReact } from "react"
 import { format} from "date-fns"
 import type { DateRange } from "react-day-picker"
 
@@ -14,14 +14,18 @@ import { ExpensesInsights } from "@/components/transactions/insights/expenses-in
 import { AccountsActions } from "@/components/accounts/accounts-actions"
 import { useTags } from "@/hooks/use-tags"
 import { useAccounts } from "@/hooks/use-accounts"
+import { useReports } from "@/hooks/use-reports"
+import type { Tag } from "@/types/tags"
 
 export function Transactions({ title, accountId, showInsights = false }: { title: string, accountId?: number, showInsights?: boolean }) {
 
-  const [filters, setFilters] = useState<TransactionFilterProps>({accountId: accountId} as TransactionFilterProps)
+  const [filters, setFilters] = useState<TransactionFilterProps>(accountId ? {accountId} : {} as TransactionFilterProps)
   const { transactions, refreshTransactions } = useTransactions()
   const { balance, refreshBalance } = useBalance()
   const { accounts, refreshAccounts } = useAccounts()
   const { tags } = useTags()
+  const { asyncFetchReportTransactions, asyncFetchReportTrends, expenseTagReport, expenseTrendReport } = useReports()
+  const [isLoading, startTransition] = useTransitionReact()
 
   useEffect(() => {
     refreshAccounts()
@@ -29,8 +33,22 @@ export function Transactions({ title, accountId, showInsights = false }: { title
   }, [filters])
 
   const asyncRefreshTransactions = async () => {
-    await refreshTransactions(filters)
-    await refreshBalance(filters)
+    startTransition(async () => {
+      await refreshTransactions(filters)
+      await refreshBalance(filters)
+      if (showInsights) {
+        await asyncFetchReportTransactions({
+          ...filters,
+          periodType: filters.periodType || 'current_month',
+          reportType: 'group_by_tag',
+        })
+        await asyncFetchReportTrends({
+          ...filters,
+          periodType: filters.periodType || 'current_month',
+          reportType: 'group_by_tag',
+        })
+      }
+    })
   }
 
   const handleDateChange = (date: DateRange) => {
@@ -43,6 +61,16 @@ export function Transactions({ title, accountId, showInsights = false }: { title
         periodType: 'custom'
       })
     }
+  }
+
+  const addTagFilter = (tag: Tag) => {
+    setFilters({ ...filters, tagId: tag.id })
+  }
+
+  const removeTagFilter = () => {
+    const newFilters = { ...filters }
+    delete newFilters.tagId
+    setFilters(newFilters)
   }
   
   return (
@@ -70,7 +98,15 @@ export function Transactions({ title, accountId, showInsights = false }: { title
         </div>
       </div>
       <BalanceInfo balance={balance} />
-      {showInsights && <ExpensesInsights filters={filters} setFilters={setFilters} />}
+      {showInsights && 
+        <ExpensesInsights
+          isLoading={isLoading}
+          expenseTagReport={expenseTagReport}
+          expenseTrendReport={expenseTrendReport}
+          addTagFilter={addTagFilter}
+          removeTagFilter={removeTagFilter}
+        />
+      }
       <div className="px-4 lg:px-6">
         <DataTable
           data={transactions}
